Added colors to terminal output
Colorama module is used here to make the output of the terminal colored.
In this file color.red is used to showcase warning or error messages to the user and color.green represents the successfullness of the funciton.
Moreover, color.white is used to showcase normal texts or input texts.