main.py is integrated with essential fun to login
main.py can now operate using automation.py and get login credentials
and then try to login.
If login is successfull it return successfull otherwise it prompts to check if the login credentials are correct.

If they are not correct, user is promoted to change them on point or else they can change it whenever they want.